working process = 97%
