+++ api/utils/serializers.py
@@ -27,4 +27,18 @@ class SendMailSerializer(serializers.ModelSerializer):
         body = data.get('body')
         return Mail.send_mail(email=email,subject=subject,body=body)
 
+class MassMailSerializer(serializers.ModelSerializer):
+    subject = serializers.CharField(source = 'Mail.subject')
+    body = serializers.CharField(source = 'Mail.body') 
+    
+    class Meta(object):
+        model = User
+        fields = (
+                'subject',
+                'body',
+                'is_active',
+                'is_student',
+                'is_teacher',
+                'is_moderator',
+                'is_admin',)
